_utils.js - moved more non-calendar-specific functions into this file ac-calendar - added prmoises to calendar building functions to group them main function calls together in sequence
