+++ ac-assets/ac-calendar.js
@@ -8,7 +8,7 @@ Use			  : All javascript calendar functions
 */
 
 // import utils
-import { createEl, addClass, debounce, addStyles } from "./_utils.js";
+import { createEl, addClass, debounce, addStyles, isMobile } from "./_utils.js";
 
 // define current url
 // this gets the current directory and adds the path to the ajax file.
@@ -49,7 +49,7 @@ const acDateEnd = acCal.getAttribute("ac-dateEnd")
 
 //
 let direction = "today"; // start calendar on current month
-let isMobile = false; // initiate as false
+
 let acStartDate = ""; // empty for initial load, will be overwritten once calendar has loaded
 let fieldDateStart = "";
 let fieldDateEnd = "";
@@ -115,7 +115,7 @@ function showSpinner(show) {
   const paramsString = `lang=${acLang}`;
   const searchParams = new URLSearchParams(paramsString);
   const response = await fetch(urlSettings + searchParams);
-  console.log(response);
+  //console.log(response);
   if (!response.ok) {
     return displayError({
       msg: "Unable to find calendar settings file",
@@ -129,11 +129,20 @@ function showSpinner(show) {
     }
 
     // write styles to document head
-    if (!errors) defineSettings();
+    if (!errors) setUpCalendar();
   }
 })();
 
-function defineSettings() {
+// use promises to define and create setting, header etc.  Finally load the calendar data
+async function setUpCalendar() {
+  await defineSettings();
+  await renderHeader();
+  await buildCalendarWrapper();
+  await addNavControls();
+  loadCal(direction);
+}
+
+async function defineSettings() {
   // define texts & settings
   txtToday = settings.texts["today"];
   txtBack = settings.texts["back"];
@@ -142,12 +151,10 @@ function defineSettings() {
   txtDateEendKO = settings.texts["end_before_start"];
   txtDatesNotAvailable = settings.texts["dates_not_available"];
   minNightsAllowed = settings.min_nights;
-
-  renderHeader();
 }
 
 // create header, fetch main colors from db and include style sheet
-function renderHeader() {
+async function renderHeader() {
   const head = document.getElementsByTagName("HEAD")[0];
   let styles = settings.styles;
 
@@ -176,16 +183,13 @@ function renderHeader() {
       metaViewPort.setAttribute("name", "viewport");
       metaViewPort.content =
         "width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0";
-      $head.appendChild(metaViewPort);
+      head.appendChild(metaViewPort);
     }
   }
-
-  // load calendar once we have the styles etc.
-  buildCalendarWrapper();
 }
 
 // create wrapper to hold calendar months
-function buildCalendarWrapper() {
+async function buildCalendarWrapper() {
   // calendar wrapper on parent page
   acWrapper = document.querySelector(`#${acWrapperID}`);
 
@@ -238,15 +242,10 @@ function buildCalendarWrapper() {
 
   // define number of months to show according to screen width - DO WE NEED TO DO THIS EVERY TIME ????
   acNumMonthsElTmp = monthsToShow();
-
-  // add nav controls to newly created nav eleemtns
-  addNavControls();
-  // load calendar
-  loadCal(direction);
 }
 
-// add calendar "back" and "next" button events
-function addNavControls() {
+// add calendar nav
+async function addNavControls() {
   var acControls = document.querySelectorAll("[data-direction]");
   acControls.forEach((btn) => {
     btn.onclick = function () {
@@ -265,6 +264,108 @@ function monthsToShow() {
     return Math.floor(acWidth / acMonthWidth);
   } else {
     return acNumMonthsInital; // user defined
+  }
+}
+
+// FETCH calendar JSON data
+function loadCal(direction) {
+  // show spinner
+  showSpinner(true);
+
+  // define data to send via fetch POST
+  let params = {
+    id_item: "" + acItemID + "",
+    lang: "" + acLang + "",
+    numMonths: "" + acNumMonthsElTmp + "",
+    startDate: "" + acStartDate + "",
+    direction: "" + direction + "",
+  };
+  //console.log(acStartDate);
+  const searchParams = new URLSearchParams(params);
+  (async function () {
+    let response = await fetch(urlCal + searchParams);
+    // If the call failed, throw an error
+    if (!response.ok) {
+      return displayError({
+        msg: "Unable to read month data",
+        code: "0.04",
+      });
+    } else {
+      // Otherwise, get the post JSON
+      const data = await response.json();
+      // define new start date for next cal load
+      acStartDate = data["start-date"];
+
+      drawCal(data);
+    }
+  })();
+}
+
+// add JSON months returned to calendar container
+function drawCal(data) {
+  if (data.error) {
+    return displayError(data.error);
+  }
+  // clear calendar wrapper contents
+  acNumMonthsEl.innerHTML = "";
+
+  // define data parts returned
+  const resMonths = data["months"];
+  const resWeekDays = data["weekdays"];
+
+  // loop through each month returmed to create calendar month
+  for (let i = 0; i < resMonths.length; i++) {
+    // month title
+    let monthTitleClone = monthTitleEl.cloneNode(true);
+    monthTitleClone.textContent = resMonths[i].month_title;
+
+    const weekDayTitlesClone = weekDayTitlesEl.cloneNode(true);
+    let weekDaysNumbersClone = weekDaysNumbersEl.cloneNode(true);
+    let days = resMonths[i].days;
+    for (let j = 0; j < days.length; j++) {
+      // clone li element
+      let dateNum = weekDayEl.cloneNode(true);
+
+      // define vars from result
+      let dateNumClass = days[j].c;
+      let dateNumFormat = days[j].df;
+      let dateNumTitle = days[j].ds;
+      let dateNumState = days[j].s;
+
+      if (dateNumState) {
+        // add date state if defined
+        dateNumTitle += " - " + dateNumState;
+      }
+      dateNum.setAttribute("title", dateNumTitle);
+      dateNum.setAttribute("id", "date_" + dateNumFormat);
+      dateNum.setAttribute("data-date", dateNumFormat);
+      // add date classes
+      for (let cl = 0; cl < dateNumClass.length; cl++) {
+        dateNum.classList.add("" + dateNumClass[cl] + "");
+      }
+      //dateNum.classList.add(""+dateNumClass+"");
+      dateNum.textContent = days[j].n;
+
+      // add date to ul
+      weekDaysNumbersClone.appendChild(dateNum);
+    }
+
+    // put month together
+    const newMonthCl = newMonthEl.cloneNode(true);
+    newMonthCl.append(
+      monthTitleClone,
+      weekDayTitlesClone,
+      weekDaysNumbersClone
+    );
+
+    // add calendar month
+    acNumMonthsEl.appendChild(newMonthCl);
+
+    // remove spinner
+    showSpinner(false);
+
+    // add click event to dates
+    activateDates();
   }
 }
 
@@ -428,124 +529,6 @@ function highlightDates() {
     }
   }
 }
-// FETCH calendar JSON data
-function loadCal(direction) {
-  // show spinner
-  showSpinner(true);
-
-  // define data to send via fetch POST
-  let params = {
-    id_item: "" + acItemID + "",
-    lang: "" + acLang + "",
-    numMonths: "" + acNumMonthsElTmp + "",
-    startDate: "" + acStartDate + "",
-    direction: "" + direction + "",
-  };
-  console.log(acStartDate);
-  const searchParams = new URLSearchParams(params);
-  (async function () {
-    let response = await fetch(urlCal + searchParams);
-    // If the call failed, throw an error
-    if (!response.ok) {
-      return displayError({
-        msg: "Unable to read month data",
-        code: "0.04",
-      });
-    } else {
-      // Otherwise, get the post JSON
-      const data = await response.json();
-      // define new start date for next cal load
-      acStartDate = data["start-date"];
-
-      drawCal(data);
-    }
-  })();
-}
-
-// add JSON months returned to calendar container
-function drawCal(data) {
-  if (data.error) {
-    return displayError(data.error);
-  }
-  // clear calendar wrapper contents
-  acNumMonthsEl.innerHTML = "";
-
-  // define data parts returned
-  const resMonths = data["months"];
-  const resWeekDays = data["weekdays"];
-
-  // loop through each month returmed to create calendar month
-  for (let i = 0; i < resMonths.length; i++) {
-    // month title
-    let monthTitleClone = monthTitleEl.cloneNode(true);
-    monthTitleClone.textContent = resMonths[i].month_title;
-
-    const weekDayTitlesClone = weekDayTitlesEl.cloneNode(true);
-    let weekDaysNumbersClone = weekDaysNumbersEl.cloneNode(true);
-    let days = resMonths[i].days;
-    for (let j = 0; j < days.length; j++) {
-      // clone li element
-      let dateNum = weekDayEl.cloneNode(true);
-
-      // define vars from result
-      let dateNumClass = days[j].c;
-      let dateNumFormat = days[j].df;
-      let dateNumTitle = days[j].ds;
-      let dateNumState = days[j].s;
-
-      if (dateNumState) {
-        // add date state if defined
-        dateNumTitle += " - " + dateNumState;
-      }
-      dateNum.setAttribute("title", dateNumTitle);
-      dateNum.setAttribute("id", "date_" + dateNumFormat);
-      dateNum.setAttribute("data-date", dateNumFormat);
-      // add date classes
-      for (let cl = 0; cl < dateNumClass.length; cl++) {
-        dateNum.classList.add("" + dateNumClass[cl] + "");
-      }
-      //dateNum.classList.add(""+dateNumClass+"");
-      dateNum.textContent = days[j].n;
-
-      // add date to ul
-      weekDaysNumbersClone.appendChild(dateNum);
-    }
-
-    // put month together
-    const newMonthCl = newMonthEl.cloneNode(true);
-    newMonthCl.append(
-      monthTitleClone,
-      weekDayTitlesClone,
-      weekDaysNumbersClone
-    );
-
-    // add calendar month
-    acNumMonthsEl.appendChild(newMonthCl);
-
-    // remove spinner
-    showSpinner(false);
-
-    // add click event to dates
-    activateDates();
-  }
-}
-
-// device detection - we are only interested if it is a mobile device or not
-// NOTE - NOT returning iPad as mobile device as it is now pretending to be a desktop agent
-function deviceDetect() {
-  if (
-    /(android|bb\d+|meego).+mobile|avantgo|bada\/|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od)|ipad|iris|kindle|Android|Silk|lge |maemo|midp|mmp|netfront|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\/|plucker|pocket|psp|series(4|6)0|symbian|treo|up\.(browser|link)|vodafone|wap|windows (ce|phone)|xda|xiino/i.test(
-      navigator.userAgent
-    ) ||
-    /1207|6310|6590|3gso|4thp|50[1-6]i|770s|802s|a wa|abac|ac(er|oo|s\-)|ai(ko|rn)|al(av|ca|co)|amoi|an(ex|ny|yw)|aptu|ar(ch|go)|as(te|us)|attw|au(di|\-m|r |s )|avan|be(ck|ll|nq)|bi(lb|rd)|bl(ac|az)|br(e|v)w|bumb|bw\-(n|u)|c55\/|capi|ccwa|cdm\-|cell|chtm|cldc|cmd\-|co(mp|nd)|craw|da(it|ll|ng)|dbte|dc\-s|devi|dica|dmob|do(c|p)o|ds(12|\-d)|el(49|ai)|em(l2|ul)|er(ic|k0)|esl8|ez([4-7]0|os|wa|ze)|fetc|fly(\-|_)|g1 u|g560|gene|gf\-5|g\-mo|go(\.w|od)|gr(ad|un)|haie|hcit|hd\-(m|p|t)|hei\-|hi(pt|ta)|hp( i|ip)|hs\-c|ht(c(\-| |_|a|g|p|s|t)|tp)|hu(aw|tc)|i\-(20|go|ma)|i230|iac( |\-|\/)|ibro|idea|ig01|ikom|im1k|inno|ipaq|iris|ja(t|v)a|jbro|jemu|jigs|kddi|keji|kgt( |\/)|klon|kpt |kwc\-|kyo(c|k)|le(no|xi)|lg( g|\/(k|l|u)|50|54|\-[a-w])|libw|lynx|m1\-w|m3ga|m50\/|ma(te|ui|xo)|mc(01|21|ca)|m\-cr|me(rc|ri)|mi(o8|oa|ts)|mmef|mo(01|02|bi|de|do|t(\-| |o|v)|zz)|mt(50|p1|v )|mwbp|mywa|n10[0-2]|n20[2-3]|n30(0|2)|n50(0|2|5)|n7(0(0|1)|10)|ne((c|m)\-|on|tf|wf|wg|wt)|nok(6|i)|nzph|o2im|op(ti|wv)|oran|owg1|p800|pan(a|d|t)|pdxg|pg(13|\-([1-8]|c))|phil|pire|pl(ay|uc)|pn\-2|po(ck|rt|se)|prox|psio|pt\-g|qa\-a|qc(07|12|21|32|60|\-[2-7]|i\-)|qtek|r380|r600|raks|rim9|ro(ve|zo)|s55\/|sa(ge|ma|mm|ms|ny|va)|sc(01|h\-|oo|p\-)|sdk\/|se(c(\-|0|1)|47|mc|nd|ri)|sgh\-|shar|sie(\-|m)|sk\-0|sl(45|id)|sm(al|ar|b3|it|t5)|so(ft|ny)|sp(01|h\-|v\-|v )|sy(01|mb)|t2(18|50)|t6(00|10|18)|ta(gt|lk)|tcl\-|tdg\-|tel(i|m)|tim\-|t\-mo|to(pl|sh)|ts(70|m\-|m3|m5)|tx\-9|up(\.b|g1|si)|utst|v400|v750|veri|vi(rg|te)|vk(40|5[0-3]|\-v)|vm40|voda|vulc|vx(52|53|60|61|70|80|81|83|85|98)|w3c(\-| )|webc|whit|wi(g |nc|nw)|wmlb|wonu|x700|yas\-|your|zeto|zte\-/i.test(
-      navigator.userAgent.substr(0, 4)
-    )
-  ) {
-    return true;
-  } else {
-    return false;
-  }
-}
 
 // once document has loaded
 document.addEventListener("DOMContentLoaded", function () {
@@ -564,9 +547,6 @@ document.addEventListener("DOMContentLoaded", function () {
     });
   }
 
-  // check if mobile
-  isMobile = deviceDetect();
-
   // detect window resize and call reloadOnResize function
   window.addEventListener("resize", debounce(reloadOnResize, 150));
 });
@@ -574,7 +554,7 @@ document.addEventListener("DOMContentLoaded", function () {
 function reloadOnResize() {
   // NOT on mobiles as they resize automatically on scroll to remove the header bar
   if (!isMobile) {
-    // recalculate number of months we can show accoring to window size
+    // recalculate number of months we can show according to window size
     acNumMonthsElTmp = monthsToShow();
     // initiate calendar
     loadCal("current");
