+++ ac-docs/error-codes.md
@@ -7,4 +7,5 @@
 <li>0.04: unable to read month data url - check calendar url</li>
 
 <li>1.01: unable to connect to database - check database user data
+
 <li>2.01: calendar install script still present - remove once script has been installed
